requirements.txt
# Core dependencies
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
altair>=5.0.0
plotly>=5.18.0
matplotlib>=3.7.0

# Machine learning
scikit-learn>=1.3.0
umap-learn>=0.5.4

# Data processing
tqdm>=4.66.0

wordcloud>=1.8.2
pillow>=9.0.0
requests>=2.28.0
python-dateutil>=2.8.2
openai>=0.27.0 